refactor: extract tree-ops + introduce shared-state module
Move dismissChildren out of the worldViewImportController monolith
into a dedicated wvImportTreeOpsController file. Introduce
wvImportSharedState.ts to own the undoEntries Map plus the
ImportStateSnapshot, SuggestionSnapshot, and UndoEntry types — these
are consumed by multiple domain controllers (tree-ops here, plus
flatten and hierarchy in PR-15/PR-16) so a shared module avoids
cross-controller imports.

Pure refactor — no behavior change, byte-identical function body
and type definitions. Continues the spine domain-split (ADR-0009).
